Overhauls application to use new Ratatouille.App behaviour
I got rid of components in Ratatouille and replaced them with
an App behaviour and runtime based on the Elm Architecture.

Unifies all the event handling in the new toby/app.ex and converts
the components to pure view functions.

There were some deeper issues with the component strategy that needed to
be solved in order to add new features, so I started looking at
alternatives. One issue is that stateful components here were only
one-level deep (they themselves only used stateless components), and it
wasn't clear how to pass events through multiple levels.

The TEA approach uses a single event handler for the whole application,
which is much simpler in comparison. The caveat is that the TEA update
function eventually gets very long, but there are strategies in the
community for dealing with that.