Adds basic node selection and shows stats for selected node
Now all the calls in Data.Provider fetch the statistics on the
selected node instead of on the local one.

For now, selecting another node requires that it already be connected.

For example, in one tab, start toby:

    $ elixir --sname foo -S mix run --no-halt

In another tab:

    $ iex --sname bar
    iex(bar@localhost)1> Node.connect(:foo@localhost)
    true

Afterwards it's possible to select the node within the node selection
overlay by moving the cursor and pressing enter.