Adds ability to drill down into an application supervision tree
When scrolling through applications on the "Applications" tab, it's now
possible to scroll through processes in the application's supervision
tree and view information about each process. This should be very
helpful when trying to understand how a given application works, as it's
easy to quickly pull up the details on a process and gain insights into
a process's purpose and relationship to other processes in the tree.

With an application selected, press `l` or `→` (right arrow key) to move
into the application's process tree pane. From there, by scrolling
vertically, each node in the tree can be selected to see process
information (as on the "Processes" tab). Press `h` or `←` to return to
the application selection.

This makes some adjustments to the cursor model to accomodate x and y
cursors, and in the case of applications, multiple y cursors.